Finish breaking up octodns.record
